This commit fixes a long standing bug in trunk - selections now persist
2008-03-04  Gabriel Burt  <gabriel.burt@gmail.com>

	This commit fixes a long standing bug in trunk - selections now persist
	across source reloads (assuming the tracks are still in the source/not
	filtered).

	* src/Core/Banshee.Services/Banshee.Playlist/AbstractPlaylistSource.cs:
	* src/Core/Banshee.Services/Banshee.Collection.Database/TrackListDatabaseModel.cs:
	Add a bool CachesJoinTableEntries property so Smart Playlists can use the
	join table but fill the CoreCache with TrackIDS not EntryIDS - needed,
	since their EntryIDs don't last long b/c they are wiped/created anew every
	time its Reloaded.

	* src/Core/Banshee.Services/Banshee.Collection/BansheeListModel.cs: Don't
	clear the selection on reload.

	* src/Core/Banshee.Services/Banshee.SmartPlaylist/SmartPlaylistSource.cs:
	Override CachesJoinTableEntries to false.

	* src/Core/Banshee.Services/Banshee.Playlist/PlaylistSource.cs:
	* src/Core/Banshee.Services/Banshee.Collection.Database/AlbumListDatabaseModel.cs:
	* src/Core/Banshee.Services/Banshee.Collection.Database/ArtistListDatabaseModel.cs:
	* src/Core/Banshee.Services/Banshee.Sources/DatabaseSource.cs: Check
	model.CachesJoinTableEntries instead of JoinTable != null.

	* src/Core/Banshee.ThickClient/Banshee.Gui/TrackActions.cs: Whitespace.

	* src/Libraries/Hyena/Hyena.Data.Sqlite/ICacheableDatabaseModel.cs:
	Inherit from ISelectable and add CachesJoinTableEntries property. 

	* src/Libraries/Hyena/Hyena.Data.Sqlite/SqliteModelCache.cs: SaveSelection
	before a Reload and RestoreSelection afterward.

	* src/Libraries/Hyena/Hyena.Data/ModelCache.cs:
	* src/Libraries/Hyena/Hyena.Data/ArrayModelCache.cs:
	* src/Libraries/Hyena/Hyena.Data/DictionaryModelCache.cs: Update to use
	Int64s where appropriate.

	* src/Libraries/Hyena/Hyena.Data/IListModel.cs: Inherit from ISelectable.

	* src/Libraries/Hyena/Hyena.Data/ModelSelection.cs: Comment out
	unimplemented IndexOf method.

	* src/Libraries/Hyena/Makefile.am: New ISelectable interface.

svn path=/trunk/banshee/; revision=3379

+++ src/Libraries/Hyena/Hyena.Data/ArrayModelCache.cs
@@ -33,21 +33,21 @@ namespace Hyena.Data
     public abstract class ArrayModelCache<T> : ModelCache<T>
     {
         protected T [] cache;
-        protected int offset = -1;
-        protected int limit = 0;
+        protected long offset = -1;
+        protected long limit = 0;
 
         public ArrayModelCache (ICacheableModel model) : base (model)
         {
             cache = new T [model.FetchCount];
         }
 
-        public override bool ContainsKey (int i)
+        public override bool ContainsKey (long i)
         {
             return (i >= offset &&
                     i <= (offset + limit));
         }
 
-        public override void Add (int i, T item)
+        public override void Add (long i, T item)
         {
             if (cache.Length != model.FetchCount) {
                 cache = new T [model.FetchCount];
@@ -63,7 +63,7 @@ namespace Hyena.Data
             limit++;
         }
 
-        public override T this [int i] {
+        public override T this [long i] {
             get { return cache [i - offset]; }
         }
